requirements.txt
streamlit
pandas
numpy
scikit-learn
rdkit-pypi